@@ -2,21 +2,21 @@ name: regular_integration_tests
 publish_to: none
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ^3.3.1
+  flutter: ">=3.19.3"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  build_runner: ^2.4.2
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
-  mockito: 5.4.1
+  mockito: ^5.4.4
   url_launcher_platform_interface: 
     git: https://github.com/kumsumit/url_launcher_platform_interface.git
   url_launcher_web:
     path: ../
+  web: ^0.5.0
