@@ -3,11 +3,11 @@ publish_to: none
 description: Web platform implementation of url_launcher
 repository: https://github.com/flutter/packages/tree/main/packages/url_launcher/url_launcher_web
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+url_launcher%22
-version: 2.0.20
+version: 2.3.0
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ^3.3.1
+  flutter: ">=3.19.3"
 
 flutter:
   plugin:
@@ -24,6 +24,7 @@ dependencies:
     sdk: flutter
   url_launcher_platform_interface: 
     git: https://github.com/kumsumit/url_launcher_platform_interface.git
+  web: ^0.5.0
 
 dev_dependencies:
   flutter_test:
