@@ -3,8 +3,8 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:html' as html;
-import 'dart:js_util';
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 import 'dart:ui_web' as ui_web;
 
 import 'package:flutter/foundation.dart';
@@ -13,6 +13,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:url_launcher_platform_interface/link.dart';
+import 'package:web/web.dart' as html;
 
 /// The unique identifier for the view type to be used for link platform views.
 const String linkViewType = '__url_launcher::link';
@@ -104,7 +105,11 @@ class LinkViewController extends PlatformViewController {
     if (_instances.isEmpty) {
       // This is the first controller being created, attach the global click
       // listener.
-      _clickSubscription = html.window.onClick.listen(_onGlobalClick);
+
+      _clickSubscription =
+          const html.EventStreamProvider<html.MouseEvent>('click')
+              .forTarget(html.window)
+              .listen(_onGlobalClick);
     }
     _instances[viewId] = this;
   }
@@ -164,11 +169,11 @@ class LinkViewController extends PlatformViewController {
   @override
   final int viewId;
 
-  late html.Element _element;
+  late html.HTMLElement _element;
 
   Future<void> _initialize() async {
-    _element = html.Element.tag('a');
-    setProperty(_element, linkViewIdProperty, viewId);
+    _element = html.document.createElement('a') as html.HTMLElement;
+    _element[linkViewIdProperty] = viewId.toJS;
     _element.style
       ..opacity = '0'
       ..display = 'block'
@@ -279,11 +284,7 @@ class LinkViewController extends PlatformViewController {
 int? getViewIdFromTarget(html.Event event) {
   final html.Element? linkElement = getLinkElementFromTarget(event);
   if (linkElement != null) {
-    // TODO(stuartmorgan): Remove this ignore (and change to getProperty<int>)
-    // once the templated version is available on stable. On master (2.8) this
-    // is already not necessary.
-    // ignore: return_of_invalid_type
-    return getProperty(linkElement, linkViewIdProperty);
+    return linkElement.getProperty<JSNumber>(linkViewIdProperty.toJS).toDartInt;
   }
   return null;
 }
@@ -312,5 +313,5 @@ html.Element? getLinkElementFromTarget(html.Event event) {
 bool isLinkElement(html.Element? element) {
   return element != null &&
       element.tagName == 'A' &&
-      hasProperty(element, linkViewIdProperty);
+      element.hasProperty(linkViewIdProperty.toJS).toDart;
 }
